Finalized (most of) Residential/Commerical functionality
The main functionality of the game (IE, incrementing and decrementing) has been implemented for residential and commercial buildings; for either zoning type, an increment button adds buildings of random sizes to one of six size categories, and a decrement button (for each size category) decrements from said categories. A slider is included to increment/decrement in steps greater than 1.