Pushing a more generalized zoning simulator (and civic simulator)
After thinking about it, I've decided to remove all the code related to the two zoning simulators (the ResSim and CommSim) in favor of a generalized class that's inherited into a simpler implementation of both, instead of having their constituent classes and constants be standalone; apart from a few lines of code being changed in the UI-related scripts, this leaves the rest of the game unchanged.

This also features a generalized civic simulator that will be inherited into a HealthSim and EduSim; hopefully implementation of either one will take a short amount of time compared to before.