Experimented with a tab system and pushing the (successful) results
I added a tab system where clicking a tab switches to some different UI panel; this will be useful for showing more stuff on the screen and organizing the different controls (IE, separating the zoning and civic simulators).

This, however, means that a few things are out of place at the moment as things get shuffled around.