Redid residential demand so that it's dependent on civic performance
Previously, residential demand was based on how many commercial job openings there were, but that implementation usually hit a wall at some point; this time, residential demand is based on the overall performance of civic buildings (performances are on a scale on the interval [0, 1]), where the cumulative performance (IE, the sum of all civic service performances) is summed up and multiplied with a fraction of the overall population. This implementation makes city growth dependent on how well the city provides for its citizens.

Apart from some balance issues, the game is functionally complete for the purposes of senior project, with the only exception being the savefile system.